Exception Handling done for user/layer

+++ src/main/java/projeto/projetoinformatico/service/LayerService.java
@@ -1,8 +1,7 @@
-package projeto.projetoinformatico.service;// LayerService.java
+package projeto.projetoinformatico.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +19,7 @@ public class LayerService {
 
     private final LayersRepository layersRepository;
     private final SearchService searchService;
+
     @Autowired
     public LayerService(LayersRepository layersRepository, SearchService searchService) {
         this.layersRepository = layersRepository;
@@ -27,14 +27,19 @@ public class LayerService {
     }
 
     public Layer createLayer(String username, LayerRequest layerRequest) {
+        if (layerRequest == null) {
+            throw new IllegalArgumentException("Layer request cannot be null");
+        }
+
         Layer newLayer = new Layer();
         newLayer.setUsername(username);
         newLayer.setLayerName(layerRequest.getName());
         newLayer.setDescription(layerRequest.getDescription());
         newLayer.setQuery(layerRequest.getQuery());
-        // Execute the SPARQL query and get the results
+
         return getLayer(layerRequest, newLayer);
     }
+
     public List<Layer> getAllLayers() {
         return layersRepository.findAll();
     }
@@ -43,22 +48,17 @@ public class LayerService {
         return layersRepository.findById(id).orElse(null);
     }
 
-
-    // TO CHECK
     public Layer updateLayer(Long id, LayerRequest layerRequest) {
         Optional<Layer> optionalLayer = layersRepository.findById(id);
         if (optionalLayer.isPresent()) {
             Layer existingLayer = optionalLayer.get();
-            // Update the existing layer with the new data
             BeanUtils.copyProperties(layerRequest, existingLayer, "id");
             existingLayer.setTimestamp(new Date());
-            // Execute the updated SPARQL query and update the search result
             return getLayer(layerRequest, existingLayer);
         }
-        return null; // or throw exception if needed
+        return null;
     }
 
-    @NotNull
     private Layer getLayer(LayerRequest layerRequest, Layer existingLayer) {
         SearchResult searchResult = searchService.executeSparqlQuery(layerRequest.getQuery());
         ObjectMapper objectMapper = new ObjectMapper();
@@ -66,12 +66,16 @@ public class LayerService {
             String searchResultJson = objectMapper.writeValueAsString(searchResult);
             existingLayer.setSearchResult(searchResultJson);
         } catch (JsonProcessingException e) {
-            e.printStackTrace(); // Handle the exception as needed
+            // Handle the exception
+            e.printStackTrace();
         }
         return layersRepository.save(existingLayer);
     }
 
     public boolean deleteLayer(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Layer ID cannot be null");
+        }
         if (layersRepository.existsById(id)) {
             layersRepository.deleteById(id);
             return true;
